Fix console errors when resizing window after leaving page
Graph continues to run (need to look at that) after navving away from the page, so the graph's container element is no longer in the DOM during the attempt to alert it to the window resize event. Am just ensuring a 0 height/width return val for the SVG dimensions here to avoid the null reference in devtools for now, but eventually need to ensure we are cleaning up all references and event handlers nicely when the graph is destroyed.